@@ -1,14 +1,4 @@
 package be.jago.ssehttp1.joke;
 
-public class ChuckNorrisJoke {
-
-    private String value;
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
+public record ChuckNorrisJoke(String value) {
 }
