@@ -13,6 +13,7 @@ import java.time.Duration;
 public class JokeController {
 
     private final JokeService jokeService;
+    private final Duration interval = Duration.ofSeconds(5);
 
     JokeController(JokeService jokeService){
         this.jokeService = jokeService;
@@ -20,7 +21,7 @@ public class JokeController {
 
     @GetMapping("/chuck-norris-joke-stream")
     public Flux<ServerSentEvent<ChuckNorrisJoke>> streamJokes() {
-        return Flux.interval(Duration.ofSeconds(5))
+        return Flux.interval(interval)
                 .map(sequence -> ServerSentEvent.<ChuckNorrisJoke>builder()
                 .data(jokeService.getRandomChuckNorrisJoke())
                 .build());
